feat relationships can now be indirect
